[JBIDE-25700] move oc binary classes to their own package
