+++ plugins/org.jboss.tools.openshift.core/META-INF/MANIFEST.MF
@@ -64,6 +64,11 @@ Export-Package: org.jboss.tools.openshift.core,
    org.jboss.tools.openshift.js,
    org.jboss.tools.openshift.test,
    org.jboss.tools.openshift.reddeer",
+ org.jboss.tools.openshift.internal.core.ocbinary;
+   x-friends:="org.jboss.tools.openshift.ui,
+   org.jboss.tools.openshift.js,
+   org.jboss.tools.openshift.test,
+   org.jboss.tools.openshift.reddeer",
  org.jboss.tools.openshift.internal.core.portforwarding;
   x-friends:="org.jboss.tools.openshift.ui,
    org.jboss.tools.openshift.js,
@@ -75,6 +80,11 @@ Export-Package: org.jboss.tools.openshift.core,
    org.jboss.tools.openshift.test,
    org.jboss.tools.openshift.reddeer",
  org.jboss.tools.openshift.internal.core.server.debug;
+  x-friends:="org.jboss.tools.openshift.ui,
+   org.jboss.tools.openshift.js,
+   org.jboss.tools.openshift.test,
+   org.jboss.tools.openshift.reddeer",
+ org.jboss.tools.openshift.internal.core.server.resources;
   x-friends:="org.jboss.tools.openshift.ui,
    org.jboss.tools.openshift.js,
    org.jboss.tools.openshift.test,
