+++ plugins/org.jboss.tools.openshift.core/src/org/jboss/tools/openshift/core/preferences/OpenShiftCorePreferences.java
@@ -19,8 +19,8 @@ import org.apache.commons.lang.StringUtils;
 import org.eclipse.osgi.util.NLS;
 import org.jboss.tools.openshift.internal.common.core.preferences.StringPreferenceValue;
 import org.jboss.tools.openshift.internal.common.core.preferences.StringsPreferenceValue;
-import org.jboss.tools.openshift.internal.core.OCBinary;
 import org.jboss.tools.openshift.internal.core.OpenShiftCoreActivator;
+import org.jboss.tools.openshift.internal.core.ocbinary.OCBinary;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
