+++ plugins/org.jboss.tools.openshift.core/src/org/jboss/tools/openshift/internal/core/ocbinary/OCBinaryValidationJob.java
@@ -8,7 +8,7 @@
  * Contributors: 
  * Red Hat, Inc. - initial API and implementation 
  ******************************************************************************/
-package org.jboss.tools.openshift.internal.core;
+package org.jboss.tools.openshift.internal.core.ocbinary;
 
 import java.util.Collection;
 
@@ -17,6 +17,7 @@ import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.Status;
 import org.eclipse.core.runtime.jobs.Job;
 import org.jboss.tools.foundation.core.plugin.log.StatusFactory;
+import org.jboss.tools.openshift.internal.core.OpenShiftCoreActivator;
 
 public class OCBinaryValidationJob extends Job {
