+++ tests/org.jboss.tools.openshift.test/src/org/jboss/tools/openshift/test/internal/core/preferences/OCBinaryValidatorTest.java
@@ -29,8 +29,8 @@ import java.util.Optional;
 import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.NullProgressMonitor;
 import org.eclipse.core.runtime.Platform;
-import org.jboss.tools.openshift.internal.core.IOCBinaryRequirement;
-import org.jboss.tools.openshift.internal.core.OCBinaryValidator;
+import org.jboss.tools.openshift.internal.core.ocbinary.IOCBinaryRequirement;
+import org.jboss.tools.openshift.internal.core.ocbinary.OCBinaryValidator;
 import org.junit.Before;
 import org.junit.Test;
 import org.osgi.framework.Version;
